fix: resolve some todo
