+++ .classpath
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry kind="src" path="duv.client.cmdline/src/main/java"/>
+	<classpathentry kind="src" path="duv.client.cmdline/src/test/java"/>
+	<classpathentry kind="src" path="duv.server/src/main/java"/>
+	<classpathentry kind="src" path="duv.server/src/test/java"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
