+++ duv.client.cmdline/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>client</name>
+	<name>duv.dap.client</name>
 	<comment></comment>
 	<projects>
 	</projects>
