+++ duv.client.cmdline/.settings/org.eclipse.m2e.core.prefs
@@ -1,4 +1,4 @@
-activeProfiles=
+activeProfiles=dap.client-0.0.1-SNAPSHOT
 eclipse.preferences.version=1
 resolveWorkspaceProjects=true
 version=1
