+++ duv.client.cmdline/pom.xml
@@ -4,7 +4,7 @@
 
     <groupId>dap</groupId>
     <artifactId>client</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <version>1</version>
     <packaging>jar</packaging>
 
     <name>client</name>
