+++ duv.client.cmdline/src/main/java/dap/client/Config.java
@@ -9,9 +9,20 @@ package dap.client;
 public final class Config {
 
     /**
-     * stcok the instance of config.
+     * stock the instance of config.
      */
     private static Config config;
+
+    /**
+     * rootUrl, default = "http://localhost:8080".
+     */
+    private String rootUrl = "http://localhost:8080";
+
+    /**
+     * @param rootUrl
+     */
+    private Config() {
+    }
 
     /**
      * singleton.
@@ -27,22 +38,10 @@ public final class Config {
     }
 
     /**
-     * rootUrl, default = "http://localhost:8080".
-     */
-    private String rootUrl = "http://localhost:8080";
-
-    /**
      * @return the rootUrl
      */
     public String getRootUrl() {
         return rootUrl;
-    }
-
-    /**
-     * @param rootUrl
-     */
-    //TODO duv by Djer Attention à l'ordre : Constantes, attributs, constructeurs, méthodes métier, getter/setter
-    private Config() {
     }
 
     /**
