+++ duv.server/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>dap</name>
+	<name>duv.dap.server</name>
 	<comment></comment>
 	<projects>
 	</projects>
