+++ duv.server/src/main/java/fr/ynov/dap/dap/Config.java
@@ -5,8 +5,52 @@ package fr.ynov.dap.dap;
  * @author David_tepoche
  *
  */
-//TODO duv by Djer Généralement on évite les "chaine magiques", et on les places dans des constantes
 public class Config {
+
+    /**
+     * name of the application.
+     */
+    private static final String APPLICATION_NAME = "YnovDaP";
+
+    /**
+     * folder's name of the credentials.
+     */
+    private static final String CREDENTIALS_FOLDER_NAME = "token";
+
+    /**
+     * file's name of the client' secret.
+     */
+    private static final String SECRETS_CLIENT_FILE_NAME = "credentials.json";
+
+    /**
+     * root of the redirect url.
+     */
+    private static final String ROOT_URL_REDIRECT = "/oAuth2Callback";
+
+    /**
+     * store the oAuth2CallbackUrl.
+     */
+    private final String oAuth2CallbackUrl = ROOT_URL_REDIRECT;
+
+    /**
+    *
+    */
+    private final String clientSecretFile = SECRETS_CLIENT_FILE_NAME;
+
+    /**
+    *
+    */
+    private final String applicationName = APPLICATION_NAME;
+
+    /**
+    *
+    */
+    private final String credentialsFolder = CREDENTIALS_FOLDER_NAME;
+
+    /**
+     * get the directory of the dataStore default = the home directory of the user.
+     */
+    private String dataStoreDirectory = System.getProperty("user.home");
 
     /**
      *
@@ -14,38 +58,6 @@ public class Config {
     public Config() {
 
     }
-
-    /**
-     * store the oAuth2CallbackUrl.
-     */
-    private String oAuth2CallbackUrl = "/oAuth2Callback";
-
-    /**
-    *
-    */
-    private String clientSecretFile = "credentials.json";
-
-    /**
-    *
-    */
-    private String applicationName = "YnovDaP";
-
-    /**
-    *
-    */
-    private String credentialsFolder = "token";
-
-    /**
-     * @return the oAuth2CallbackUrl
-     */
-    public String getoAuth2CallbackUrl() {
-        return oAuth2CallbackUrl;
-    }
-
-    /**
-     * get the directory of the dataStore default = the home directory of the user.
-     */
-    private String dataStoreDirectory = System.getProperty("user.home");
 
     /**
      *
@@ -56,24 +68,10 @@ public class Config {
     }
 
     /**
-     * @return the dataStoreDirectory
+     * @return the applicationName
      */
-    public String getDataStoreDirectory() {
-        return this.dataStoreDirectory;
-    }
-
-    /**
-     * @param dataDirectory the dataStoreDirectory to set
-     */
-    public void setDataStoreDirectory(final String dataDirectory) {
-        this.dataStoreDirectory = dataDirectory;
-    }
-
-    /**
-     * @return the credentialsFolder
-     */
-    public String getCredentialsFolder() {
-        return credentialsFolder;
+    public String getApplicationName() {
+        return applicationName;
     }
 
     /**
@@ -84,10 +82,31 @@ public class Config {
     }
 
     /**
-     * @return the applicationName
+     * @return the credentialsFolder
      */
-    public String getApplicationName() {
-        return applicationName;
+    public String getCredentialsFolder() {
+        return credentialsFolder;
+    }
+
+    /**
+     * @return the dataStoreDirectory
+     */
+    public String getDataStoreDirectory() {
+        return this.dataStoreDirectory;
+    }
+
+    /**
+     * @return the oAuth2CallbackUrl
+     */
+    public String getoAuth2CallbackUrl() {
+        return oAuth2CallbackUrl;
+    }
+
+    /**
+     * @param dataDirectory the dataStoreDirectory to set
+     */
+    public void setDataStoreDirectory(final String dataDirectory) {
+        this.dataStoreDirectory = dataDirectory;
     }
 
 }
