+++ duv.server/src/main/java/fr/ynov/dap/dap/google/CalendarService.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.security.GeneralSecurityException;
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport;
@@ -13,34 +12,11 @@ import com.google.api.services.calendar.Calendar;
 import com.google.api.services.calendar.model.Event;
 import com.google.api.services.calendar.model.Events;
 
-import fr.ynov.dap.dap.Config;
-
 /**
  * @author David_tepoche
  */
 @Service
 public class CalendarService extends BaseService {
-    /**
-     * link config.
-     */
-    @Autowired
-    // TODO duv by Djer Tu pourrais utiliser la config du parent (en ajoutant un
-    // getter protected)
-    private Config config;
-
-    /**
-     *
-     * @param userId user key
-     * @return Gmail service.
-     * @throws IOException              throw by the getCredential from the
-     *                                  baseService
-     * @throws GeneralSecurityException throw by the getCredential from the
-     *                                  baseService
-     */
-    private Calendar getService(final String userId) throws GeneralSecurityException, IOException {
-        return new Calendar.Builder(GoogleNetHttpTransport.newTrustedTransport(), JACKSON_FACTORY,
-                getCredential(userId)).setApplicationName(config.getApplicationName()).build();
-    }
 
     @Override
     protected final String getClassName() {
@@ -60,13 +36,27 @@ public class CalendarService extends BaseService {
     public List<Event> getLastEvent(final Integer nbrOfResult, final String user)
             throws IOException, GeneralSecurityException {
 
-        DateTime now = new DateTime(System.currentTimeMillis());
+        final DateTime now = new DateTime(System.currentTimeMillis());
 
         Events events;
 
         events = getService(user).events().list("primary").setMaxResults(nbrOfResult).setTimeMin(now)
                 .setOrderBy("startTime").setSingleEvents(true).execute();
-        List<Event> items = events.getItems();
+        final List<Event> items = events.getItems();
         return items;
+    }
+
+    /**
+     *
+     * @param userId user key
+     * @return Gmail service.
+     * @throws IOException              throw by the getCredential from the
+     *                                  baseService
+     * @throws GeneralSecurityException throw by the getCredential from the
+     *                                  baseService
+     */
+    private Calendar getService(final String userId) throws GeneralSecurityException, IOException {
+        return new Calendar.Builder(GoogleNetHttpTransport.newTrustedTransport(), JACKSON_FACTORY,
+                getCredential(userId)).setApplicationName(getConfig().getApplicationName()).build();
     }
 }
